requirements.txt
google-generativeai
python-dotenv
click
